Desktop/겨울 프로젝트/DPT-chatbot-main/requirements.txt
django
langchain
langchain-community
langchain-core
langchain-openai
langchain-google-genai
faiss-cpu
openai
pymupdf
python-dotenv